% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddGeneAnnotation.R
\name{AddGeneAnnotation}
\alias{AddGeneAnnotation}
\title{Add Gene Annotation to TENET Object}
\usage{
AddGeneAnnotation(obj, species = "human", version = "hg38")
}
\arguments{
\item{obj}{A TENET object.}

\item{species}{The species for which the annotation is required. Default is \code{"human"}.}

\item{version}{The genome version (e.g., \code{"hg38"}, \code{"hg19"}, \code{"mm10"}). Default is \code{"hg38"}.}
}
\value{
The TENET object with updated \code{feature_list$gtf}.
}
\description{
This function reads a gene annotation file from the package's \verb{inst/bash/}
directory and adds it to the \code{feature_list$gtf} field of the TENET object.
}
