% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TENET_RunPipeline_old.R
\name{runTENETPlus_old}
\alias{runTENETPlus_old}
\title{Run TENET+ Pipeline}
\usage{
runTENETPlus_old(
  obj,
  outdir = ".",
  ncores = 2,
  historyLength = 1,
  species = NULL,
  mode = 1,
  bashFile = NULL
)
}
\arguments{
\item{obj}{A TENET object.}

\item{outdir}{Directory to store intermediate and result files.}

\item{ncores}{Number of parallel jobs.}

\item{historyLength}{Parameter for history length.}

\item{species}{Species to use (if NULL, uses obj@input$metadata$Species).}

\item{mode}{Mode to run (as defined in TENET_Plus_for_py.sh).}

\item{bashFile}{Path to the bash script. If using an internal script, retrieve its path via system.file().}
}
\value{
Updated TENET object with output stored in obj@output.
}
\description{
Executes the TENET+ pipeline by writing out the necessary input files,
invoking an external bash script (and Python scripts), and reading the results
back into the TENET object.
}
