% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TENET_Conversion.R
\name{SeuratToTENET}
\alias{SeuratToTENET}
\title{Convert Seurat Object to TENET Object}
\usage{
SeuratToTENET(
  seurat_obj,
  Species = "human",
  DEG = list(),
  DAR = list(),
  TF_list = list(),
  pseudotime = "pseudotime",
  cell_select = "cell_select"
)
}
\arguments{
\item{seurat_obj}{A Seurat object.}

\item{Species}{A character string indicating the species.}

\item{DEG}{A list or data frame for DEG information.}

\item{DAR}{A list or data frame for DAR information.}

\item{TF_list}{A list or data frame for TF_list information.}

\item{pseudotime}{Column name or data frame with pseudotime values.}

\item{cell_select}{Column name or data frame indicating cell selection.}
}
\value{
A TENET object.
}
\description{
Converts a Seurat object (with RNA and optional ATAC assays) into a TENET object.
}
