% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TENET_Methods.R
\name{show,TENET-method}
\alias{show,TENET-method}
\title{Show Method for TENET Objects}
\usage{
\S4method{show}{TENET}(object)
}
\arguments{
\item{object}{A TENET object.}
}
\description{
Custom display of a TENET object in the R console.
}
