% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TENET_Class.R
\docType{class}
\name{TENET}
\alias{TENET}
\title{TENET Class Definition}
\description{
Defines the S4 class TENET with slots for input and output.
}
\section{Slots}{

\describe{
\item{\code{input}}{A list containing RNA/ATAC counts, metadata, and feature list.}

\item{\code{output}}{A list to store processed data/results.}
}}

